moved frontend api roles to framework bundle
