+++ packages/framework/src/Resources/translations/messages.cs.po
@@ -88,6 +88,12 @@ msgstr "<strong><a href=\"{{ url }}\">Notifikační lišta</a></strong> byla ús
 msgid "<strong><a href=\"{{ url }}\">SEO category</a></strong> has been saved"
 msgstr "<strong><a href=\"{{ url }}\">SEO kombinace kategorie</a></strong> byla uložena"
 
+msgid "Access to all the complaints created under the user's company"
+msgstr "Přístup ke všem reklamacím vytvořeným pod firmou uživatele"
+
+msgid "Access to all the orders created under the user's company"
+msgstr "Přístup ke všem objednávkám vytvořeným pod firmou uživatele"
+
 msgid "Accessories"
 msgstr "Příslušenství"
 
@@ -394,6 +400,9 @@ msgstr "Průměrné trvání všech cronů za posledních sedm dní"
 msgid "Avg duration (mm:ss)"
 msgstr "Prům. trvání (mm:ss)"
 
+msgid "B2B data and user management"
+msgstr "Správa B2B údajů a uživatelů"
+
 msgid "Back to category selection"
 msgstr "Zpět na volbu kategorie"
 
@@ -583,6 +592,9 @@ msgstr "Kanonická URL"
 msgid "Cardholder canceled the payment"
 msgstr "Držitel platební karty zrušil platbu"
 
+msgid "Cart manipulation and order creation"
+msgstr "Manipulace s košíkem a vytváření objednávky"
+
 msgid "Catalog number"
 msgstr "Katalogové číslo"
 
@@ -750,6 +762,9 @@ msgstr "Reklamace č."
 
 msgid "Complaint Nr. <strong><a href=\"{{ url }}\">{{ number }}</a></strong> modified"
 msgstr "Byla upravena reklamace č. <strong><a href=\"{{ url }}\">{{ number }}</a></strong>"
+
+msgid "Complaint creation"
+msgstr "Vytváření reklamace"
 
 msgid "Complaint detail URL"
 msgstr "URL detailu reklamace"
@@ -1068,6 +1083,12 @@ msgstr "Jméno zákazníka"
 
 msgid "Customer phone number"
 msgstr "Telefonní číslo zákazníka"
+
+msgid "Customer sees prices"
+msgstr "Zákazník vidí ceny"
+
+msgid "Customer self manage"
+msgstr "Spravovat své údaje"
 
 msgid "Customer street"
 msgstr "Ulice zákazníka"
