+++ packages/frontend-api/src/Resources/translations/messages.en.po
@@ -4,28 +4,7 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Language: en\n"
 
-msgid "Access to all the complaints created under the user's company"
-msgstr ""
-
-msgid "Access to all the orders created under the user's company"
-msgstr ""
-
-msgid "B2B data and user management"
-msgstr ""
-
-msgid "Cart manipulation and order creation"
-msgstr ""
-
-msgid "Complaint creation"
-msgstr ""
-
 msgid "Customer not found."
-msgstr ""
-
-msgid "Customer sees prices"
-msgstr ""
-
-msgid "Customer self manage"
 msgstr ""
 
 msgid "{0}Customer has not logged in yet.|[1,Inf[ Customer uses login via %loginTypes%."
