+++ upgrade-notes/backend_20251113_081933.md
@@ -1,0 +1,6 @@
+#### moved frontend api roles to framework bundle ([#4269](https://github.com/shopsys/shopsys/pull/4269))
+
+- classes were moved from the frontend api package to the framework package
+    - `Shopsys\FrontendApiBundle\Component\Security\FrontendApiRoleHierarchyProvider` -> `Shopsys\FrameworkBundle\Model\Customer\User\Role\CustomerUserRoleHierarchyProvider`
+    - `Shopsys\FrontendApiBundle\Component\Security\FrontendApiRoleProvider` -> `Shopsys\FrameworkBundle\Model\Customer\User\Role\CustomerUserRoleProvider`
+    - warning: this is a potencial BC break, if you used the classes directly!
